\name{getValuesBlock_enhanced}
\alias{getValuesBlock_enhanced}
\title{getValuesBlock_enhanced
Easier-to-use function for grabbing data out of a Raster*.}
\usage{
  getValuesBlock_enhanced(x, r1, r2, c1, c2,
    format = "array", ...)
}
\arguments{
  \item{x}{TODO.}

  \item{r1}{TODO.}

  \item{r2}{TODO.}

  \item{c1}{TODO.}

  \item{c2}{TODO.}

  \item{format}{TODO.}

  \item{...}{TODO.}
}
\value{
  An array.
}
\description{
  getValuesBlock_enhanced Easier-to-use function for
  grabbing data out of a Raster*.
}
\examples{
# TODO
}
\author{
  Jonathan A. Greenberg
}
\seealso{
  \code{\link[raster]{getValues}}
}

