% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tahoe_highrez_training_data.R
\docType{data}
\name{tahoe_highrez_training}
\alias{tahoe_highrez_training}
\title{Point and polygon files for use with spatial.tools}
\description{
Point and polygon files for use with spatial.tools
}
\examples{
library("raster")
library("rgdal")
tahoe_highrez_training_polygons <- readOGR(
	dsn=system.file("external", package="spatial.tools"),layer="tahoe_highrez_training")
spplot(tahoe_highrez_training_polygons,zcol="Class")
tahoe_highrez_training_points <- readOGR(
	dsn=system.file("external", package="spatial.tools"),layer="tahoe_highrez_training_points")
spplot(tahoe_highrez_training_points,zcol="SPECIES")
tahoe_highrez_training_points_utm <- readOGR(
	dsn=system.file("external", package="spatial.tools"),
	layer="tahoe_highrez_training_points_utm")
print(projection(tahoe_highrez_training_points_utm))
}
\author{
Jonathan A. Greenberg \email{spatial.tools@estarcion.net}
}
\keyword{data}
